+++ coloraide/spaces/lab/__init__.py
@@ -72,7 +72,7 @@ class Lab(Labish, Space):
 
         super().__init__(**kwargs)
         order = alg.order(round(self.channels[self.indexes()[0]].high, 5))
-        self.achromatic_threshold = (1 * 10.0 ** order) / 1_000_000
+        self.achromatic_threshold = max((1 * 10.0 ** order) / 1_000_000, 1e-12)
 
     def is_achromatic(self, coords: Vector) -> bool:
         """Check if color is achromatic."""